Jwt auth

+++ app/routers/forms_router.py
@@ -1,8 +1,9 @@
 from enum import Enum
 from typing import List
 
-from app.schemas import FormResponses
+from app.schemas import FormResponses, User
 from app.services import forms as crud
+from auth.auth_router import get_current_auth_user
 from database import get_db
 from fastapi import APIRouter, Depends, HTTPException
 from app import schemas
@@ -14,13 +15,13 @@ router = APIRouter(
 )
 
 
-async def is_authorized(db, form_id, user_id):
+async def is_authorized(db, form_id, user_id) -> True or Exception:
     form = await crud.get_form(db, form_id=form_id)
     if form is None:
         raise HTTPException(status_code=404, detail="Form not found")
-    if form.creator_id is not user_id:
+    if form.creator_id != user_id:
         raise HTTPException(status_code=401, detail="Unauthorized")
-    return form
+    return True
 
 
 @router.get("/{form_id}", response_model=schemas.FormGet)
@@ -32,7 +33,9 @@ async def get_form(form_id: int, db: AsyncSession = Depends(get_db)):
 
 
 @router.post("", response_model=schemas.FormGet)
-async def create_form(form: schemas.FormCreate, creator_id: int, db: AsyncSession = Depends(get_db)):
+async def create_form(form: schemas.FormCreate, db: AsyncSession = Depends(get_db),
+                      user: User = Depends(get_current_auth_user)):
+    creator_id = user.id
     new_form = await crud.create_form(db=db, form=form, creator_id=creator_id)
     return new_form
 
@@ -44,13 +47,18 @@ class TypeField(str, Enum):
 
 
 @router.post("/{form_id}/fields", response_model=schemas.FormFieldGet)
-async def add_field(field: schemas.FormFieldCreate, field_type: TypeField, form_id: int, db: AsyncSession = Depends(get_db)):
-    form = await crud.get_form(db, form_id=form_id)
-    return await crud.create_field(db=db, field=field, form_id=form_id, field_type=field_type)
+async def add_field(field: schemas.FormFieldCreate, field_type: TypeField, form_id: int,
+                    db: AsyncSession = Depends(get_db),
+                    user: User = Depends(get_current_auth_user)):
+    if field_type == TypeField.text:
+        field.options = []
+    if await is_authorized(db, form_id=form_id, user_id=user.id):
+        return await crud.create_field(db=db, field=field, form_id=form_id, field_type=field_type)
 
 
 @router.post("/{form_id}/responses", response_model=schemas.FormResponseGet)
-async def create_form_response(answers: List[schemas.FormAnswerCreate], form_id: int, db: AsyncSession = Depends(get_db)):
+async def create_form_response(answers: List[schemas.FormAnswerCreate], form_id: int,
+                               db: AsyncSession = Depends(get_db)):
     form = await crud.get_form(db, form_id=form_id)
     if form is None:
         raise HTTPException(status_code=404, detail="Form not found")
@@ -61,34 +69,31 @@ async def create_form_response(answers: List[schemas.FormAnswerCreate], form_id:
 
 
 @router.patch("/{form_id}/publish", response_model=schemas.FormGet)
-async def publish_form(form_id: int, user_id: int, db: AsyncSession = Depends(get_db)):
-    form = await is_authorized(db, user_id=user_id, form_id=form_id)
-    await crud.publish_form(db, form_id=form_id)
-    await db.refresh(form)
-    return form
+async def publish_form(form_id: int, db: AsyncSession = Depends(get_db),
+                       user: User = Depends(get_current_auth_user)):
+    if await is_authorized(db, user_id=user.id, form_id=form_id):
+        await crud.publish_form(db, form_id=form_id)
+        return await crud.get_form(db, form_id=form_id)
 
 
 @router.patch("/{form_id}/unpublish", response_model=schemas.FormGet)
-async def unpublish_form(form_id: int, user_id: int, db: AsyncSession = Depends(get_db)):
-    form = await is_authorized(db, user_id=user_id, form_id=form_id)
-    if form is None:
-        raise HTTPException(status_code=404, detail="Form not found")
-    if form.creator_id is not user_id:
-        raise HTTPException(status_code=401, detail="Unauthorized")
-    await crud.unpublish_form(db, form_id=form_id)
-    await db.refresh(form)
-    return form
+async def unpublish_form(form_id: int, db: AsyncSession = Depends(get_db),
+                         user: User = Depends(get_current_auth_user)):
+    if await is_authorized(db, user_id=user.id, form_id=form_id):
+        await crud.unpublish_form(db, form_id=form_id)
+        return await crud.get_form(db, form_id=form_id)
 
 
 @router.delete("/{form_id}/delete", response_model=schemas.FormGet)
-async def delete_form(form_id: int, user_id: int, db: AsyncSession = Depends(get_db)):
-    form = await is_authorized(db, form_id=form_id, user_id=user_id)
-    await crud.delete_form(db, form_id=form_id)
-    return form
+async def delete_form(form_id: int, db: AsyncSession = Depends(get_db),
+                      user: User = Depends(get_current_auth_user)):
+    if await is_authorized(db, form_id=form_id, user_id=user.id):
+        return await crud.delete_form(db, form_id=form_id)
 
 
 @router.get("/{form_id}/answers", response_model=List[FormResponses])
-async def get_responses(form_id: int, user_id: int, db: AsyncSession = Depends(get_db)):
-    await is_authorized(db, form_id=form_id, user_id=user_id)
-    responses = await crud.get_responses(db, form_id=form_id)
-    return responses
+async def get_responses(form_id: int, db: AsyncSession = Depends(get_db),
+                        user: User = Depends(get_current_auth_user)):
+    if await is_authorized(db, form_id=form_id, user_id=user.id):
+        responses = await crud.get_responses(db, form_id=form_id)
+        return responses
